feat: add test infrastructure for new document types and block codec
This commit adds build system support for testing material, decal, and asset
document modules, along with the scene block codec functionality. New test
runner targets are defined for each module, with source dependencies properly
linked to enable comprehensive unit testing coverage.

+++ src/scene_document_internal.h
@@ -126,7 +126,7 @@ SceneRepairReplaceResult scene_document_internal_replace_surface_material(
     int map_x,
     int map_y,
     SceneSurfaceKind surface,
-    uint8_t replacement_material_id,
+    uint16_t replacement_material_id,
     DocumentStateId resulting_state
 );
